+++ src/Interfaces/Filters.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Task interface.
+ * Filters interface.
  *
  * @category Interface
  * @package  Awsm\WP_Wrapper\Building_Plans
@@ -9,22 +9,20 @@
  * @link     https://awesome.ug
  */
 
-namespace Awsm\WP_Wrapper\Building_Plans;
+namespace Awsm\WP_Wrapper\Interfaces;
 
 /**
- * Task Plugin_Interface
+ * Interface Hooks_Interface.
  *
- * @package Awsm\WP_Wrapper\Building_Plans
+ * @package SvenWagener\WP_Plugin
  *
  * @since 1.0.0
  */
-interface Task {
+interface Filters {
 	/**
-	 * Register function which executes scripts.
+	 * Adding Filters.
 	 *
 	 * @since 1.0.0
-	 *
-	 * @return mixed
 	 */
-	public function run();
+	public function add_filters();
 }
